filesys.rename: add debug logging and retries (see below)
Without the retries, this stuff does not work as expected:
If the win32 transacted "atomic" rename finds that there is already some other
transacted rename running for the target file, it returns ret=6800. In the
original code this meant falling back to another (pseudo-atomic) rename, which
then likely had the same problem, causing another fallback to the non-atomic
rename. It shouldn't work like that.

After this changeset, if it doesn't succeed with a rename, it sleeps 1ms and
then retries SAME KIND of rename. It retries that up to 100 times and then
would fall back to another rename method (this usually does not happen).

Tested on Windows 2000 Server and Windows Server 2008 R2.