fix edit-log migration issue with wrong line endings
while trying a migration from linux-based 1.5.x wiki to windows-based 1.9.3,
i had strange failures at 1.8.9999 step, crashed in .refresh() function
while parsing the edit-log. Often this crash happens when edit-log was already
damaged before, but in this case it seem due to some offset miscalculation
in the MoinMoin.logfile code which was maybe triggered by the migration code
converting \n line endings to \r\n.

a fresh edit-log created on windows just has \n line endings, so we better
use binary mode in the converter when opening the output edit-log.

i also fixed another inconsistency in MoinMoin.logfile - when it opens the
logfile for output, it used mode 'a', but for reading it is 'rb'.
changed to 'ab' for consistency.

--HG--
branch : editlogfixes