getSubscribers: cleanups, optimizations, locking fix, .valid fix
removed superfluous/wrong cache.lock('r') call, the cache file is already
closed and unlocked. And as it is not used again in this method, there is
no need to lock it again. This is the same state as for the other code path,
if the cache existed.

fix: check the user.valid state, so behaviour of cached and uncached function
is the same.

docstrings/comments updated/fixed.

typos, formatting, style fixed.

use same name for the timer as for the function

break out of search loop as soon as we know that user IS subscribed