graphingwiki.scripts.gwiki-editpage: Fixed path handling collab: Added a patch to fix importing wiki configs only from the config directory.
NB: The multiconfig.py.patch works also with non-collab wikis, but only if the configuration is as recommended in HelpOnInstallation and the example MoinMoin server configuration - i.e. the configuration directory is inserted to the beginning of sys.path. I couldn't think of a general solution for discovering non-standard config paths from sys.path.